backend/requirements.txt
fastapi==0.113.0
uvicorn==0.30.6
pillow==10.4.0
numpy==2.1.1
scikit-learn==1.5.1
joblib==1.4.2
python-multipart==0.0.9
opencv-python==4.10.0.84
scikit-image==0.24.0
#tensorflow==2.17.0